Replace "list" variable names
Several method used a variable named "list". "list" is also a type and a function name in Python; using "list" as a variable name is not forbidden in Python, but isn't good practice and could lead to confusion and unexpected behavior, especially if later the list() function was attempted to be used in one of the methods that had a "list" variable. This commit replaces all "list" variable names with "volume_list" or "workflow_list" as appropriate for the context.