fix APFS ram disk mount command and sleep
diskuitl mount /dev/diskXX does not work for APFS volumes.

diskutil erasedisk TYPE VOLUMENAME DISKPATH works for both APFS and HFS+. This command properly initializes/erases/mounts the disk.

Sleep 2 seconds, because even though the process is finished for initializing the disk, it's actually not fully initialized. Prior to this, gurl was being invoked 2 tenths of a second after the subprocess was finished, resulting in disk IO errors.